# openapi.yaml
---
openapi: 3.0.3

info:
  title: embedding-server API
  version: 0.1.0
  description: |
    マルチモーダル埋め込み用のジョブ管理 API。
    テキストは同期、画像・multimodal は非同期（202 + job id / webhook / GET）。
    同一テキストの埋め込み結果はサーバー内部でキャッシュされる。
    公開 API は API_KEY、/internal は INTERNAL_API_KEY を Bearer token として送る。

servers:
  - url: https://embeddings.mumumu6.net
    description: 公開環境
  - url: http://localhost:8080
    description: 開発環境

security:
  - ExternalBearerAuth: []

tags:
  - name: embeddings
    description: API_KEY で認証し、テキスト・画像・テキスト＋画像から埋め込みジョブを作成する
  - name: internal-worker
    description: INTERNAL_API_KEY で認証するワーカー用

paths:
  /v1/embeddings/text:
    post:
      operationId: postEmbeddingsText
      tags: [embeddings]
      summary: テキストの埋め込みベクトルを返す
      description: |
        ジョブを DB に作成し、処理完了まで待機してから埋め込みベクトルを返す。
        同一テキストはサーバー内部でキャッシュされ、キャッシュヒット時は即時返却。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  $ref: "#/components/schemas/EmbeddingText"
      responses:
        "200":
          description: 埋め込みベクトル
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbeddingResult"
        "400":
          description: テキストが無効、または JSON が不正です
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          description: ジョブを受け付けられません。しばらく待ってから再送してください
          headers:
            Retry-After:
              description: 推奨リトライ待機秒数
              required: true
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "504":
          description: ジョブ処理がタイムアウトしました
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/embeddings/images:
    post:
      operationId: postEmbeddingsImages
      tags: [embeddings]
      summary: 画像群の埋め込みジョブを受け付ける
      description: |
        ジョブを DB に作成してすぐ job id を返す（非同期）。
        完了・失敗時は webhook_url があれば、WebhookNotification の本文で id と status のみを通知する。
        状態・結果は認証付き GET /v1/embeddings/jobs/{id} で取得する。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [images]
              properties:
                images:
                  type: array
                  minItems: 1
                  maxItems: 4
                  items:
                    $ref: "#/components/schemas/EmbeddingImageFile"
                webhook_url:
                  $ref: "#/components/schemas/WebhookUrl"
      callbacks:
        jobStatus:
          '{$request.body#/webhook_url}':
            post:
              summary: 埋め込みジョブの完了・失敗通知
              requestBody:
                required: true
                content:
                  application/json:
                    schema:
                      $ref: "#/components/schemas/WebhookNotification"
              responses:
                "204":
                  description: 通知を受信した
      responses:
        "202":
          description: ジョブを受け付けた
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbeddingJobAccepted"
        "400":
          description: multipart または images フィールドが不正です（枚数・サイズ・形式を含む）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/embeddings/multimodal:
    post:
      operationId: postEmbeddingsMultimodal
      tags: [embeddings]
      summary: テキストと画像群の埋め込みジョブを受け付ける
      description: |
        text + images 用。画像を含む場合は非同期（202 + job id）。
        text のみの場合もこの endpoint では非同期にする（同期が必要なら /v1/embeddings/text を使う）。
        完了・失敗時は webhook_url があれば、WebhookNotification の本文で id と status のみを通知する。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                text:
                  $ref: "#/components/schemas/EmbeddingText"
                images:
                  type: array
                  minItems: 1
                  maxItems: 4
                  items:
                    $ref: "#/components/schemas/EmbeddingImageFile"
                webhook_url:
                  $ref: "#/components/schemas/WebhookUrl"
              anyOf:
                - required: [text]
                - required: [images]
      callbacks:
        jobStatus:
          '{$request.body#/webhook_url}':
            post:
              summary: 埋め込みジョブの完了・失敗通知
              requestBody:
                required: true
                content:
                  application/json:
                    schema:
                      $ref: "#/components/schemas/WebhookNotification"
              responses:
                "204":
                  description: 通知を受信した
      responses:
        "202":
          description: ジョブを受け付けた
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbeddingJobAccepted"
        "400":
          description: text・images の指定が不正です（枚数・サイズ・形式を含む）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/embeddings/jobs/{id}:
    get:
      operationId: getEmbeddingsJob
      tags: [embeddings]
      summary: 埋め込みジョブの状態と結果を取得する
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: ジョブ状態
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbeddingJobStatus"
        "404":
          description: 指定したジョブが見つかりません
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /internal/worker/jobs/claim:
    post:
      security:
        - InternalBearerAuth: []
      operationId: claimWorkerJob
      tags: [internal-worker]
      summary: 次の pending ジョブをロックして取得
      description: |
        DB 内の pending ジョブを1件 processing にし、id と payload を返す。
        `kinds` で取得したい種別を指定できる。省略時は text と image の両方（text 優先）。
        text は画像なし、image は画像付き（multimodal 含む）。
        該当ジョブが無い場合は 204。
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimWorkerJobRequest"
      responses:
        "200":
          description: ジョブあり
          content:
            application/json:
              schema:
                type: object
                required: [id, payload]
                properties:
                  id:
                    $ref: "#/components/schemas/JobId"
                  payload:
                    $ref: "#/components/schemas/WorkerJobPayload"
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          description: kinds の指定が不正です
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /internal/worker/jobs/{id}/complete:
    post:
      security:
        - InternalBearerAuth: []
      operationId: completeWorkerJob
      tags: [internal-worker]
      summary: ジョブを完了にする
      parameters:
        - $ref: "#/components/parameters/JobId"
      requestBody:
        description: |
          worker が計算した埋め込み結果を返す。
          テキスト埋め込みジョブの場合、サーバーはこの結果を対応するキーでキャッシュへ保存する。
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [result]
              properties:
                result:
                  $ref: "#/components/schemas/EmbeddingResult"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          description: リクエストが不正です
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: 指定したジョブが見つかりません
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /internal/worker/jobs/{id}/fail:
    post:
      security:
        - InternalBearerAuth: []
      operationId: failWorkerJob
      tags: [internal-worker]
      summary: ジョブを失敗にする
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          description: リクエストが不正です
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: 指定したジョブが見つかりません
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    ExternalBearerAuth:
      type: http
      scheme: bearer
    InternalBearerAuth:
      type: http
      scheme: bearer

  parameters:
    JobId:
      name: id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/JobId"

  responses:
    NoContent:
      description: No Content
    Unauthorized:
      description: API key is missing or invalid
      headers:
        WWW-Authenticate:
          schema:
            type: string
          description: Bearer authentication challenge
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: エラーメッセージ

    JobId:
      type: string
      format: uuid

    WorkerJobKind:
      type: string
      enum: [text, image]
      description: text は画像なしジョブ、image は画像付きジョブ（multimodal 含む）

    ClaimWorkerJobRequest:
      type: object
      properties:
        kinds:
          type: array
          items:
            $ref: "#/components/schemas/WorkerJobKind"
          description: |
            取得したいジョブ種別。省略時は text と image（text を優先して返す）。

    EmbeddingResult:
      type: object
      required: [vector]
      properties:
        vector:
          type: array
          minItems: 4096
          maxItems: 4096
          items:
            type: number
            format: float
          description: |
            正規化済みの埋め込みベクトル。
            Qwen3-VL-Embedding-8B の出力次元は 4096。

    EmbeddingJobAccepted:
      type: object
      required: [id]
      properties:
        id:
          $ref: "#/components/schemas/JobId"

    EmbeddingJobStatus:
      type: object
      required: [id, status]
      properties:
        id:
          $ref: "#/components/schemas/JobId"
        status:
          type: string
          enum: [pending, processing, completed, failed]
        result:
          $ref: "#/components/schemas/EmbeddingResult"

    WebhookUrl:
      type: string
      format: uri
      maxLength: 2048
      description: |
        ジョブ完了・失敗時に WebhookNotification を POST する公開 HTTPS URL（任意）。
        通知を合図に、既知の job id の状態・結果を認証付き GET /v1/embeddings/jobs/{id} で取得する。

    WebhookNotification:
      type: object
      required: [id, status]
      properties:
        id:
          $ref: "#/components/schemas/JobId"
        status:
          type: string
          enum: [completed, failed]
          description: ジョブの最終状態。Webhook は完了または失敗時だけ送信する。
      example:
        id: 00000000-0000-0000-0000-000000000000
        status: completed

    EmbeddingText:
      type: string
      minLength: 1
      maxLength: 8192
      description: |
        埋め込みに用いるテキスト。
        1〜8192 文字。空文字は不可。

    EmbeddingImageFile:
      type: string
      format: binary
      maxLength: 20971520
      description: |
        埋め込みに用いる画像バイナリ。
        対応形式: image/png, image/jpeg, image/webp。
        1ファイルあたり最大 20 MiB（20971520 bytes）。

    EmbeddingImageObject:
      type: object
      required: [key]
      properties:
        key:
          type: string
          description: worker がオブジェクトストレージから取得する画像オブジェクトキー
      description: worker が参照する画像オブジェクト

    EmbeddingImageObjects:
      type: array
      minItems: 1
      items:
        $ref: "#/components/schemas/EmbeddingImageObject"
      description: worker が参照する画像オブジェクトのリスト

    WorkerJobPayload:
      type: object
      description: worker が処理するジョブ内容
      properties:
        text:
          $ref: "#/components/schemas/EmbeddingText"
        image_objects:
          $ref: "#/components/schemas/EmbeddingImageObjects"